fix(cart): make sure quantity is integer when adding item to cart, fixes #13

+++ src/lib/cart/cart.ts
@@ -24,7 +24,7 @@ export type CartType = {
 
 export const CartAdd = z.object({
     product_id: z.string().startsWith("variant_"),
-    quantity: z.number().min(1).max(99),
+    quantity: z.number().int().min(1).max(99),
 });
 
 export type CartAdd = z.infer<typeof CartAdd>;
